+++ client/src/main/java/org/realityforge/gwt/appcache/client/html5/Html5ApplicationCache.java
@@ -41,6 +41,7 @@ public final class Html5ApplicationCache
   {
     super( eventBus );
     registerListeners0();
+    enableAppCache();
   }
 
   @Nonnull